requirements.txt
spacy
pytesseract
pdf2image
PyMuPDF
Pillow
sentence-transformers
openai


